+++ src/database/initialization/has_proposal_field_value_permission.sql
@@ -71,6 +71,7 @@ BEGIN
 							= has_proposal_field_value_permission.user_keycloak_user_id
 							AND euga.user_group_keycloak_organization_id
 								= pg.grantee_keycloak_organization_id
+							AND NOT is_expired(euga.not_after)
 					)
 				)
 			)